+++ src/main/java/org/example/repository/CreateJsonData.java
@@ -1,4 +1,4 @@
-package org.example.config;
+package org.example.repository;
 
 public class CreateJsonData {
