+++ src/main/java/org/example/repository/QueryStringParser.java
@@ -1,4 +1,4 @@
-package org.example.config;
+package org.example.repository;
 
 public class QueryStringParser {
