+++ src/test/java/org/example/MainTest.java
@@ -53,6 +53,8 @@ class MainTest {
             }
 
         }
+
+
     }
 
     // 디렉토리 및 그 안의 파일을 삭제하는 메서드
